fix: comment in is-dao-or-extension, comment cleanup
The Clarity linter should not mess this up anymore but might be a regression, comment block was not before the line it changed on quite a few is-dao-or-extension functions, generally causes errors at deploy step not sure why they weren't caught here. Maybe we didn't generate a new set wince that happened?