fix: auto approve ft if deposited by tx-sender
deposit is gated to be tx-sender of either owner or agent for direct interaction with faktory dex through a wrapper. if tx-sender calls deposit for a DAO token through buying it, the DAO token gets added to the approved contract list and is listed in the print event.

based on this print event we can then check if we have the action proposal voting contract record in the DB and have the agent approve it for proposal functionality.