fix: scaffold starting criteria for generated docs
Following successful pattern with contract analysis, might be on to something useful here! End goal is developer docs that live with the contracts and user docs generated on the GitBook with clear linking between the two. Bubbles up from contract code -> dev docs -> user docs and process should be replicable.