feat: add recursive clarity decoder function
We use this across a few of our projects already so want to include it here as a utility instead. The current cvToValue implementation will return clarity values with complex objects like lists and tuples. This will go through and decode everything with some options to preserve containers and ensure json compatibility based on the underlying cvToValue() usage.

This is best used in conjunction with a type and type guard, something we can generate in the future based on the contract abi, so that we can say the object should be this shape based on the fields in the contract. But instead of { type: uint, value: 235 } we just get  ready to rock.